src/main/java/ru/netology/domain/FileOpenManager.java
package ru.netology.domain;

import java.util.*;
import java.util.Map;

public class FileOpenManager {
    Map<String, String> map = new HashMap<>();

    public void register(String name, String extension) {
        map.put(extension.toLowerCase(), name);
    }

    public String checkName(String extension) {
        return map.get(extension.toLowerCase());
    }

    public void removeConnection(String extension) {
        map.remove(extension.toLowerCase());
    }

    public List<String> getAllExtensions() {
        Set<String> strings = map.keySet();
        List<String> list = new ArrayList<>(strings);
        ApplicationsComparator comparator = new ApplicationsComparator();
        list.sort(comparator);
        return list;
    }

    public Set<String> getAllApps() {
        Collection<String> values = map.values();
        Set<String> appSet = new HashSet<>(values);
        return appSet;
    }
}